feat(ReservationTimeService): 특정 시간에 대한 예약이 존재하는 시간 삭제 시 예외 기능 구현
- 이에 따른 테스트 코드 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,8 +2,10 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
+import roomescape.exception.ExistReservationInReservationTimeException;
 import roomescape.exception.NotExistReservationTimeException;
 import roomescape.service.dto.ReservationTimeInput;
 import roomescape.service.dto.ReservationTimeOutput;
@@ -12,9 +14,11 @@ import roomescape.service.dto.ReservationTimeOutput;
 public class ReservationTimeService {
 
     private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     public ReservationTimeOutput createReservationTime(ReservationTimeInput input) {
@@ -31,6 +35,10 @@ public class ReservationTimeService {
     public void deleteReservationTime(long id) {
         ReservationTime reservationTime = reservationTimeDao.find(id)
                 .orElseThrow(() -> new NotExistReservationTimeException(String.format("%d는 없는 id 입니다.", id)));
+
+        if (reservationDao.isExistsByTimeId(id)) {
+            throw new ExistReservationInReservationTimeException(String.format("%d에 해당하는 예약이 있습니다.", id));
+        }
 
         reservationTimeDao.delete(reservationTime.getId());
     }
